Improve and streamline Ursa hello world instructions
Instructions to run the Ursa 'Hello World' test are expanded for clarity.

Additionally, paths are made relative where possible to reduce the number
of changes that need to be made. A couple of paths that were inconsistent
between steps were also updated.

Finally, a gitignore is added for Ursa to hide output products from git.